+++ ext/scripts/popup.js
@@ -5,7 +5,8 @@
 
 /**
  * Create a link to the extension Details tab.
- * This is a clone of the same function found in helpers. */
+ * This is a clone of the same function found in helpers.
+ */
 function LinkDetails() {
   const extensionId = chrome.runtime.id,
     ua = navigator.userAgent
@@ -18,7 +19,13 @@ function LinkDetails() {
   return `chrome${url}`
 }
 
-chrome.tabs.query({ currentWindow: true, active: true }, () => {
+/**
+ * LinkTos handles the behavior of the popup buttons.
+ * It retrieves website domains from local storage, displays the extension version,
+ * checks if file scheme access is allowed, and provides links to various options pages.
+ * @returns {void}
+ */
+async function LinkTos() {
   // list website domains
   chrome.storage.local.get(`settingsWebsiteDomains`, (store) => {
     const domains = store.settingsWebsiteDomains
@@ -51,19 +58,11 @@ chrome.tabs.query({ currentWindow: true, active: true }, () => {
   // link to the options page and fonts tab
   const fonts = document.getElementById(`fonts`)
   fonts.title = `Open the RetroTxt fonts options`
-  fonts.addEventListener(`click`, () => {
-    chrome.tabs.create({
-      url: `${chrome.runtime.getURL(`html/options.html`)}#top?t=fonts`,
-    })
-  })
+  fonts.addEventListener(`click`, () => TabTo(`fonts`))
   // link to the options page and display tab
   const display = document.getElementById(`display`)
   display.title = `Open the RetroTxt display options`
-  display.addEventListener(`click`, () => {
-    chrome.tabs.create({
-      url: `${chrome.runtime.getURL(`html/options.html`)}#top?t=display`,
-    })
-  })
+  display.addEventListener(`click`, () => TabTo(`display`))
   // link to the RetroTxt Extension Details tab
   const extension = document.getElementById(`extension`)
   extension.title = `Open the RetroTxt Extension Details`
@@ -75,16 +74,31 @@ chrome.tabs.query({ currentWindow: true, active: true }, () => {
   // link to the options page and settings tab
   const settings = document.getElementById(`settings`)
   settings.title = `Open the RetroTxt settings`
-  settings.addEventListener(`click`, () => {
-    chrome.tabs.create({
-      url: `${chrome.runtime.getURL(`html/options.html`)}#top?t=settings`,
-    })
-  })
+  settings.addEventListener(`click`, () => TabTo(`settings`))
   // link to the options page to add new urls
   const access = document.getElementById(`allowAccess`)
-  access.addEventListener(`click`, () => {
+  access.addEventListener(`click`, () => TabTo(`settings`))
+}
+
+/**
+ * Opens a new tab or updates an existing tab with the specified tab button.
+ * @param {string} value - The value to be passed as the `t` query parameter in the URL.
+ */
+function TabTo(value) {
+  const path = `${chrome.runtime.getURL(`html/options.html`)}`
+  chrome.tabs.query({ url: path }, (tabs) => {
+    for (let i = 0; i < tabs.length; i++) {
+      chrome.tabs.update(tabs[i].id, {
+        active: true,
+        url: `${path}#top?t=${value}`,
+      })
+      return
+    }
     chrome.tabs.create({
-      url: `${chrome.runtime.getURL(`html/options.html`)}#top?t=settings`,
+      active: true,
+      url: `${path}#top?t=${value}`,
     })
   })
-})
+}
+
+chrome.tabs.query({ currentWindow: true, active: true }, LinkTos)